% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bam.files.coverage.R
\name{bam.files.coverage.plot}
\alias{bam.files.coverage.plot}
\title{Plots a barchart of bam files coverage.}
\usage{
bam.files.coverage.plot(path = ".", intervals = c(0, 5, 10, 20, 30, 50),
  pattern = "-coverage-hist.txt$")
}
\arguments{
\item{path}{The path to the directory containing your coverage statistics file}

\item{intervals}{A numeric vector with coverage intervals}

\item{pattern}{A regular expression pattern. Only file names which match the regular expression will be returned}
}
\value{
A ggplot object to be printed
}
\description{
Plot a barchart of bam files coverage. It needs coverage statistic files generated with the command:
bedtools coverage -hist -abam sample.bam -b target_regions.bed | grep ^all > sample-coverage-hist.txt
}
\examples{
fpath <- system.file("extdata", "sample2-coverage-hist.txt", package="iSeqsR")
directory <- dirname(fpath)
bam.files.coverage.plot(path = directory, pattern="-coverage-hist.txt$", intervals = c(0,10,20,30,50))

}

