% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.data.R
\name{build.fishplot.matrix}
\alias{build.fishplot.matrix}
\title{Build a matrix ready to be passed to createFishObject}
\usage{
build.fishplot.matrix(fishplot.data.frame, clustering.results,
  relapse2 = FALSE)
}
\arguments{
\item{fishplot.data.frame}{data from data.for.fishplot funtion}

\item{clustering.results}{Clustering results from dbscan::dbscan function}

\item{relapse2}{If the data from a second relapse is present (default = FALSE)}
}
\value{
A matrix to be used in createFishObject function
}
\description{
Starting with data from data.for.fishplot and clustering results from dbscan::dbscan
build a matrix to be used in createFishObject
}
