% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{coMut.plot.main}
\alias{coMut.plot.main}
\title{Produces a very basic coMut plot.}
\usage{
coMut.plot.main(all.variants, min.occurence = 2, legend.position = "right")
}
\arguments{
\item{min.occurence}{Minimun number of patient in which the gene should be mutated to be included in the plot}

\item{legend.position}{The position where to draw the legend, "none" for no legend}

\item{variants.table}{A tbl_df obtained with read.variants.tsv function or a data frame with impact, Patient and gene_name columns}
}
\value{
A ggplot
}
\description{
Given a variants table data frame, builds a very basic coMut plot (https://www.broadinstitute.org/blog/visualizing-cancer-genome).
The plot shows mutation type at gene level. It also shows if a gene in a sample is hit by mutation with different types.
}

