% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{coMut.plot.mutations.in.sample}
\alias{coMut.plot.mutations.in.sample}
\title{##' Produces the right plot in a coMut plot: a barchart with the number of patients in which a gene.}
\usage{
coMut.plot.mutations.in.sample(all.variants, min.occurence = 2)
}
\description{
##' Produces the right plot in a coMut plot: a barchart with the number of patients in which a gene.
}

