% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{coMut.plot}
\alias{coMut.plot}
\title{Produces a coMut plot.}
\usage{
coMut.plot(all.variants, min.occurence = 2, feature = "impact")
}
\arguments{
\item{min.occurence}{Minimun number of patient in which the gene should be mutated to be included in the plot}

\item{variants.table}{A tbl_df obtained with read.variants.tsv function or a data frame with impact, Patient and gene_name columns}
}
\value{
A ggplot
}
\description{
Given a variants table data frame, builds a very basic coMut plot (https://www.broadinstitute.org/blog/visualizing-cancer-genome).
The plot shows mutation type at gene level and on the right a bar chart with the number of patients in which a gene is.
It also shows if a gene in a sample is hit by mutation with different types.
}
