% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coverage.R
\name{coverage.cumulative.distribution.plot}
\alias{coverage.cumulative.distribution.plot}
\title{Plot the cumulative distribution of the fraction of targeted bases covered by n reads}
\usage{
coverage.cumulative.distribution.plot(path = ".",
  pattern = "-coverage-hist.txt$")
}
\arguments{
\item{path}{The path to the directory containing your coverage statistics file}

\item{pattern}{A regular expression pattern. Only file names which match the regular expression will be returned}
}
\description{
Plot the cumulative distribution of the fraction of targeted bases covered by n reads. It needs coverage statistic files generated with the command:
bedtools coverage -hist -abam sample.bam -b target_regions.bed | grep ^all > sample-coverage-hist.txt
It is based on the code from this the post in Getting Things Done in Genetics: http://www.gettinggeneticsdone.com/2014/03/visualize-coverage-exome-targeted-ngs-bedtools.html
}
\examples{
fpath <- system.file("extdata", "sample2-coverage-hist.txt", package="iSeqsR")
directory <- dirname(fpath)
coverage.cumulative.distribution.plot(path = directory, pattern="-coverage-hist.txt$")

}

