% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.data.R
\name{data.for.fishplot}
\alias{data.for.fishplot}
\title{Prepares a data frame ready to be transformed to a matrix for fishplot.}
\usage{
data.for.fishplot(data, diagnosis = "C01_ES_C03_REM",
  relapse = "C02_RIC_C03_REM", relapse2 = FALSE, coverage = 30,
  min.vaf = 5)
}
\arguments{
\item{data}{The data frame with starting data}

\item{diagnosis}{The name of the diagnosis in the SAMPLE_NAME column}

\item{relapse}{The name of the relapse in the SAMPLE_NAME column}

\item{relapse2}{The name of the second relapse in the SAMPLE_NAME column (default = FALSE, second relaps).}

\item{coverage}{The minimum coverage of the called variants (default = 30)}

\item{min.vaf}{The minimum VAF value in % (dafault = 5)}

\item{relapse2}{If the data of a secondo relapse is present (default = FALSE)}
}
\value{
A data frame with these columns: VAF_DIAGNOSIS, GENE, VAF_RELAPSE
}
\description{
Prepares a data frame to be transformed to a matrix for fishplot. As imput 'data' it needs a data frame
with COVERAGE, VAF and SAMPLE_NAME columns
}
