\name{iSeqsR-package}
\alias{iSeqsR-package}
\alias{iSeqsR}
\docType{package}
\title{
\packageTitle{iSeqsR}
}
\description{
\packageDescription{iSeqsR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iSeqsR}
\packageIndices{iSeqsR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{iSeqsR}

Maintainer: \packageMaintainer{iSeqsR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
