% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.data.R
\name{matrix.for.fishplot}
\alias{matrix.for.fishplot}
\title{Get a matrix with VAF_DIAGNOSIS, VAF_RELAPSE and eventualy VAF_RELAPSE2 columns and GENES as row names to ben passed todbscan::dbscan}
\usage{
matrix.for.fishplot(data, relapse2 = FALSE)
}
\arguments{
\item{data}{The data frame obtained from data.for.fishplot function}

\item{relapse2}{If the data from a second relapse is present (default = FALSE)}
}
\value{
A mtrix with VAF_DIAGNOSIS, VAF_RELAPSE and eventually VAR_RELAPSE2 columns and GENES as row names
}
\description{
Starting from the data.frame from data.for.fishplot function
obtains a matrix with VAF_DIAGNOSIS, VAF_RELAPSE and eventually VAF_RELAPSE and GENES as row names
to be passed to dbscan::dbscan clustering function
}
