% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{merge.2.variant.tables}
\alias{merge.2.variant.tables}
\title{Merges two variants.tables.}
\usage{
\method{merge}{2.variant.tables}(t1, t2)
}
\arguments{
\item{t1}{a variants table obtained by read.variants.tsv}

\item{t2}{a variants table obtained by read.variants.tsv}
}
\value{
a data frame with variants from both t1 and t2. If a variant is present in both tables it keeps column for the first table.
}
\description{
Merges two variants tables obtained from read.variants.tsv function. It uses chromosome, genomic position, reference, alternative allele as a key.
In the case of a variant present in both tables it uses columns for the first table.
}

