% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{merge.variant.tables}
\alias{merge.variant.tables}
\title{Merges variants.tables.}
\usage{
\method{merge}{variant.tables}(variants.list)
}
\arguments{
\item{variants.list}{a  list of variants table obtained by read.variants.tsv}
}
\value{
a data frame with variants obtained by merging all variant tables in the variants.list parameter; uses chrosome, position, alternative allele and reference allele as key.
}
\description{
Merges variants tables obtained from read.variants.tsv function. It uses chromosome, genomic position, reference, alternative allele as a key.
In the case of a variant present in both tables it uses columns for the first table.
}

