% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational.spectrum.R
\name{mutations.spectrum.barchart}
\alias{mutations.spectrum.barchart}
\title{Build a barchart with mutational spectrum per patient}
\usage{
mutations.spectrum.barchart(variants.table)
}
\arguments{
\item{variants.table}{A tbl_df obtained with read.variants.tsv function or a data frame with chr, pos, ref, alt and Patient columns.}
}
\value{
A ggplot object
}
\description{
Build a barchart with mutational spectrum of six transition and transversion categories for each patient.
}

