% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bam.files.coverage.R
\name{plot.bam.files.coverage}
\alias{plot.bam.files.coverage}
\title{Plot histogram of bam files coverage}
\usage{
plot.bam.files.coverage(intervals = c(0, 5, 10, 20, 30, 50),
  pattern = "-coverage-hist.txt$")
}
\arguments{
\item{intervals}{A numeric vector with coverage intervals}

\item{pattern}{The pattern to be matched to find coverage statistics file in the current directory}
}
\description{
Plot histogram of bam files coverage
}

