% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{read.variants.tsv}
\alias{read.variants.tsv}
\title{Reads a tab separated file with variants information}
\usage{
read.variants.tsv(filename = "variants.txt", path = "")
}
\arguments{
\item{filename}{The name of the file produced by vcf_2_table.py}

\item{path}{The path to the directory containing your coverage statistics file}
}
\value{
A dplyr tbl_df with variants information
}
\description{
Reads a tab separated file with variants information produced by vcf_2_table.py script from iSeqs suite https://github.com/alexcoppe/iSeqs
Can load data from any tsv file, it only requires a column called "Patient"
}
