% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coverage.R
\name{samples.mean.coverage}
\alias{samples.mean.coverage}
\title{Calculates mean coverage for all samples in a directory}
\usage{
samples.mean.coverage(path = ".", pattern = "-coverage-hist.txt$")
}
\arguments{
\item{path}{The path to the directory containing your coverage statistics file}

\item{pattern}{A regular expression pattern. Only file names which match the regular expression will be returned}
}
\description{
Calculates mean coverage for all samples in a given directory. It needs coverage statistic files generated with the command:
bedtools coverage -hist -abam sample.bam -b target_regions.bed | grep ^all > sample-coverage-hist.txt
}
\examples{
fpath <- system.file("extdata", "sample2-coverage-hist.txt", package="iSeqsR")
directory <- dirname(fpath)
samples.mean.coverage(path = directory, pattern="-coverage-hist.txt$")

}
