% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{summarise.variants.by.patient}
\alias{summarise.variants.by.patient}
\title{Summarises the number of variants per patient}
\usage{
summarise.variants.by.patient(variants.table)
}
\arguments{
\item{variants.table}{A tbl_df obtained with read.variants.tsv function or a data frame with impact, Patient and key columns. Key column should be a column with the following string: chr,pos,alt,ref}
}
\value{
A data frame with one line for each patient. It has 3 columns: total number of variants, number of variants with MODERATE impact and number of variants with HIGH impact
}
\description{
Given variants data obtained with read.variants.tsv, returns a data frame with the number of variats found in each patient
The returned data.frame has 3 columns: total number of variants, number of variants with MODERATE predicted impact and 
number of variants with HIGH predicted impact for each patient.
}

