% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{coMut.plot.hit.genes}
\alias{coMut.plot.hit.genes}
\title{##' Produces the right plot in a coMut plot: a barchart with the number of patients in which a gene.}
\usage{
coMut.plot.hit.genes(all.variants, min.occurence = 2,
  legend.position = "right", feature = "impact")
}
\description{
##' Produces the right plot in a coMut plot: a barchart with the number of patients in which a gene.
}
