% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.table.R
\name{patients.by.number.of.variants.plot}
\alias{patients.by.number.of.variants.plot}
\title{Plots a barchart of number of variants per patient}
\usage{
patients.by.number.of.variants.plot(variants.table)
}
\arguments{
\item{variants.table}{A tbl_df obtained with read.variants.tsv function or a data frame with impact, Patient and key columns. Key column should be a column with the following string: chr,pos,alt,ref}
}
\value{
A barchart with the number of variants in each patient.
}
\description{
Given variants data obtained with read.variants.tsv, plots a barchart with the number of variants found in each patient
}

